requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
arxiv>=2.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
